# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aplikasi Gacor API Documentation",
        "description": "Documentation for the Aplikasi Gacor Package Delivery Management System",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/api/laporans": {
            "get": {
                "tags": [
                    "LaporanPengiriman"
                ],
                "summary": "Ambil semua laporan pengiriman",
                "operationId": "b5c47725fa95f25f1b449422c9afb5e6",
                "responses": {
                    "200": {
                        "description": "Data laporan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Laporan pengiriman retrieved successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_laporan": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "wilayah": {
                                                        "type": "string",
                                                        "example": "Jakarta"
                                                    },
                                                    "jumlah_paket": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "status_pengiriman": {
                                                        "type": "string",
                                                        "example": "pending"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "LaporanPengiriman"
                ],
                "summary": "Buat laporan pengiriman baru",
                "operationId": "48d0a2bde5552917ec3417e13a689a36",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "wilayah",
                                    "jumlah_paket",
                                    "status_pengiriman"
                                ],
                                "properties": {
                                    "wilayah": {
                                        "type": "string",
                                        "example": "Jakarta Selatan"
                                    },
                                    "jumlah_paket": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "status_pengiriman": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "dalam perjalanan",
                                            "selesai"
                                        ],
                                        "example": "pending"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Laporan berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Laporan pengiriman created successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "id_laporan": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "wilayah": {
                                                    "type": "string",
                                                    "example": "Jakarta Selatan"
                                                },
                                                "jumlah_paket": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "status_pengiriman": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/laporans/{id}": {
            "get": {
                "tags": [
                    "LaporanPengiriman"
                ],
                "summary": "Ambil detail laporan pengiriman",
                "operationId": "fa2945f9932d49dfbc65f7e621ba9359",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail laporan ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Laporan pengiriman retrieved successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "id_laporan": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "wilayah": {
                                                    "type": "string",
                                                    "example": "Jakarta"
                                                },
                                                "jumlah_paket": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "status_pengiriman": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Laporan tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "LaporanPengiriman"
                ],
                "summary": "Update laporan pengiriman",
                "operationId": "ea134502d6a92db16768313d8d9e6f5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "wilayah": {
                                        "type": "string",
                                        "example": "Bandung"
                                    },
                                    "jumlah_paket": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "status_pengiriman": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "dalam perjalanan",
                                            "selesai"
                                        ],
                                        "example": "dalam perjalanan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Laporan berhasil diperbarui"
                    },
                    "404": {
                        "description": "Laporan tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "LaporanPengiriman"
                ],
                "summary": "Hapus laporan pengiriman",
                "operationId": "8e9efdb8f550955d597d13da0bcb806e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Laporan berhasil dihapus"
                    },
                    "404": {
                        "description": "Laporan tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/laporans/filter/status": {
            "get": {
                "tags": [
                    "LaporanPengiriman"
                ],
                "summary": "Filter laporan berdasarkan status pengiriman",
                "operationId": "b036c6449cfe1e4045d18f8e2a7dbb88",
                "parameters": [
                    {
                        "name": "status_pengiriman",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "dalam perjalanan",
                                "selesai"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data laporan yang difilter berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/laporans/filter/wilayah": {
            "get": {
                "tags": [
                    "LaporanPengiriman"
                ],
                "summary": "Filter laporan berdasarkan wilayah",
                "operationId": "d8e9e4234f8f5459e2b9ae656e41d516",
                "parameters": [
                    {
                        "name": "wilayah",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data laporan yang difilter berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}